requirements.txt
openai
google-generativeai
keyboard
pyperclip
pystray
customtkinter
Pillow
opencv-python
ffpyplayer
numpy
